Added code seperation with different files
