+++ StockMarket_app/Modules/StockProfileDetails.js
@@ -1,0 +1,36 @@
+export default async function getStockprofileDetail() {
+     const getStockprofileDetailAPI = await fetch(`https://stocksapi-uhe1.onrender.com/api/stocks/getstocksprofiledata`);
+     const getStockprofileDetails = await getStockprofileDetailAPI.json();
+
+     for (let getStockprofileDetail in getStockprofileDetails) {
+
+          if (getStockprofileDetail != "message") {
+               getStockprofileDetails[getStockprofileDetail].map((stockProfiles) => {
+                    buildStockProfileView(stockProfiles)
+               });
+          }
+     }
+}
+
+function buildStockProfileView(stockProfiles) {
+
+     const stockDetail = document.getElementById('stock-detail');
+     const selectedStock = document.getElementById(`seletedStock`);
+
+     for (let stockProfile in stockProfiles) {
+
+          if (stockProfile != `_id` && selectedStock.value == stockProfile) {
+
+               const stockProfileHTML = `
+                    <dl>
+                         <dt>${stockProfile}</dt>
+                         <dd>
+                            ${stockProfiles[stockProfile].summary}
+                         </dd>
+                    </dl>
+
+               `
+               stockDetail.innerHTML = stockProfileHTML
+          }
+     }
+}
