--- StockMarket_app/Modules/companyValue.js
@@ -1,14 +1,0 @@
-const bookValue = {
-     "message": "stocks Data",
-     "stocksStatsData": [
-          {
-               "AAPL": {
-                    "bookValue": 3.953,
-                    "profit": 0.24493
-               },
-
-          }
-     ]
-}
-
-export default bookValue;
