--- StockMarket_app/Modules/marketData.js
@@ -1,355 +1,0 @@
-const stockData = {
-     "message": "stocks Data",
-     "stocksData": [
-          {
-               "_id": "64bf6a8cb1e39462cf026fa3",
-               "AAPL": {
-                    "5y": {
-                         "value": [
-                              54.38509750366211,
-                              54.12268829345703,
-                              52.4731559753418,
-                              42.81575393676758,
-                              37.95118713378906,
-                              40.044349670410156,
-                              41.658721923828125,
-                              45.896697998046875,
-                              48.48691940307617,
-                              42.30131912231445,
-                              48.00661087036133,
-                              51.67405700683594,
-                              50.63106918334961,
-                              54.531593322753906,
-                              60.5673942565918,
-                              65.06928253173828,
-                              71.71175384521484,
-                              75.58489227294922,
-                              66.75675201416016,
-                              62.2470817565918,
-                              71.91866302490234,
-                              77.82781982421875,
-                              89.54032135009766,
-                              104.32626342773438,
-                              126.69168853759766,
-                              113.90745544433594,
-                              107.07164001464844,
-                              117.09423065185547,
-                              130.73532104492188,
-                              130.01609802246094,
-                              119.47369384765625,
-                              120.53043365478516,
-                              129.71701049804688,
-                              122.95781707763672,
-                              135.37362670898438,
-                              144.1705322265625,
-                              150.07138061523438,
-                              140.0706024169922,
-                              148.2867431640625,
-                              163.63014221191406,
-                              176.03274536132812,
-                              173.26687622070312,
-                              163.6905059814453,
-                              173.31890869140625,
-                              156.4842987060547,
-                              147.73944091796875,
-                              135.90846252441406,
-                              161.5453643798828,
-                              156.28677368164062,
-                              137.57049560546875,
-                              152.6415252685547,
-                              147.35572814941406,
-                              129.55271911621094,
-                              143.8710174560547,
-                              146.9819793701172,
-                              164.67222595214844,
-                              169.44561767578125,
-                              177.0051727294922,
-                              193.97000122070312,
-                              193.1300048828125,
-                              192.99000549316406
-                         ],
-                         "timeStamp": [
-                              1533096000,
-                              1535774400,
-                              1538366400,
-                              1541044800,
-                              1543640400,
-                              1546318800,
-                              1548997200,
-                              1551416400,
-                              1554091200,
-                              1556683200,
-                              1559361600,
-                              1561953600,
-                              1564632000,
-                              1567310400,
-                              1569902400,
-                              1572580800,
-                              1575176400,
-                              1577854800,
-                              1580533200,
-                              1583038800,
-                              1585713600,
-                              1588305600,
-                              1590984000,
-                              1593576000,
-                              1596254400,
-                              1598932800,
-                              1601524800,
-                              1604203200,
-                              1606798800,
-                              1609477200,
-                              1612155600,
-                              1614574800,
-                              1617249600,
-                              1619841600,
-                              1622520000,
-                              1625112000,
-                              1627790400,
-                              1630468800,
-                              1633060800,
-                              1635739200,
-                              1638334800,
-                              1641013200,
-                              1643691600,
-                              1646110800,
-                              1648785600,
-                              1651377600,
-                              1654056000,
-                              1656648000,
-                              1659326400,
-                              1662004800,
-                              1664596800,
-                              1667275200,
-                              1669870800,
-                              1672549200,
-                              1675227600,
-                              1677646800,
-                              1680321600,
-                              1682913600,
-                              1685592000,
-                              1688184000,
-                              1689962125
-                         ]
-                    },
-                    "1y": {
-                         "value": [
-                              156.28677368164062,
-                              137.57049560546875,
-                              152.6415252685547,
-                              147.35572814941406,
-                              129.55271911621094,
-                              143.8710174560547,
-                              146.9819793701172,
-                              164.67222595214844,
-                              169.44561767578125,
-                              177.0051727294922,
-                              193.97000122070312,
-                              193.1300048828125,
-                              192.74929809570312
-                         ],
-                         "timeStamp": [
-                              1659326400,
-                              1662004800,
-                              1664596800,
-                              1667275200,
-                              1669870800,
-                              1672549200,
-                              1675227600,
-                              1677646800,
-                              1680321600,
-                              1682913600,
-                              1685592000,
-                              1688184000,
-                              1689963868
-                         ]
-                    },
-                    "1mo": {
-                         "value": [
-                              183.9600067138672,
-                              187,
-                              186.67999267578125,
-                              185.27000427246094,
-                              188.05999755859375,
-                              189.25,
-                              189.58999633789062,
-                              193.97000122070312,
-                              192.4600067138672,
-                              191.3300018310547,
-                              191.80999755859375,
-                              190.67999267578125,
-                              188.61000061035156,
-                              188.0800018310547,
-                              189.77000427246094,
-                              190.5399932861328,
-                              190.69000244140625,
-                              193.99000549316406,
-                              193.72999572753906,
-                              195.10000610351562,
-                              193.1300048828125,
-                              192.78990173339844
-                         ],
-                         "timeStamp": [
-                              1687354200,
-                              1687440600,
-                              1687527000,
-                              1687786200,
-                              1687872600,
-                              1687959000,
-                              1688045400,
-                              1688131800,
-                              1688391000,
-                              1688563800,
-                              1688650200,
-                              1688736600,
-                              1688995800,
-                              1689082200,
-                              1689168600,
-                              1689255000,
-                              1689341400,
-                              1689600600,
-                              1689687000,
-                              1689773400,
-                              1689859800,
-                              1689963778
-                         ]
-                    },
-                    "3mo": {
-                         "value": [
-                              164.7920684814453,
-                              165.1016387939453,
-                              163.54379272460938,
-                              163.53379821777344,
-                              168.17738342285156,
-                              169.44561767578125,
-                              169.35574340820312,
-                              168.30718994140625,
-                              167.2187042236328,
-                              165.5609893798828,
-                              173.33026123046875,
-                              173.26034545898438,
-                              171.53274536132812,
-                              173.3202667236328,
-                              173.510009765625,
-                              172.57000732421875,
-                              172.07000732421875,
-                              172.07000732421875,
-                              172.69000244140625,
-                              175.0500030517578,
-                              175.16000366210938,
-                              174.1999969482422,
-                              171.55999755859375,
-                              171.83999633789062,
-                              172.99000549316406,
-                              175.42999267578125,
-                              177.3000030517578,
-                              177.25,
-                              180.08999633789062,
-                              180.9499969482422,
-                              179.5800018310547,
-                              179.2100067138672,
-                              177.82000732421875,
-                              180.57000732421875,
-                              180.9600067138672,
-                              183.7899932861328,
-                              183.30999755859375,
-                              183.9499969482422,
-                              186.00999450683594,
-                              184.9199981689453,
-                              185.00999450683594,
-                              183.9600067138672,
-                              187,
-                              186.67999267578125,
-                              185.27000427246094,
-                              188.05999755859375,
-                              189.25,
-                              189.58999633789062,
-                              193.97000122070312,
-                              192.4600067138672,
-                              191.3300018310547,
-                              191.80999755859375,
-                              190.67999267578125,
-                              188.61000061035156,
-                              188.0800018310547,
-                              189.77000427246094,
-                              190.5399932861328,
-                              190.69000244140625,
-                              193.99000549316406,
-                              193.72999572753906,
-                              195.10000610351562,
-                              193.1300048828125,
-                              192.7949981689453
-                         ],
-                         "timeStamp": [
-                              1682083800,
-                              1682343000,
-                              1682429400,
-                              1682515800,
-                              1682602200,
-                              1682688600,
-                              1682947800,
-                              1683034200,
-                              1683120600,
-                              1683207000,
-                              1683293400,
-                              1683552600,
-                              1683639000,
-                              1683725400,
-                              1683811800,
-                              1683898200,
-                              1684157400,
-                              1684243800,
-                              1684330200,
-                              1684416600,
-                              1684503000,
-                              1684762200,
-                              1684848600,
-                              1684935000,
-                              1685021400,
-                              1685107800,
-                              1685453400,
-                              1685539800,
-                              1685626200,
-                              1685712600,
-                              1685971800,
-                              1686058200,
-                              1686144600,
-                              1686231000,
-                              1686317400,
-                              1686576600,
-                              1686663000,
-                              1686749400,
-                              1686835800,
-                              1686922200,
-                              1687267800,
-                              1687354200,
-                              1687440600,
-                              1687527000,
-                              1687786200,
-                              1687872600,
-                              1687959000,
-                              1688045400,
-                              1688131800,
-                              1688391000,
-                              1688563800,
-                              1688650200,
-                              1688736600,
-                              1688995800,
-                              1689082200,
-                              1689168600,
-                              1689255000,
-                              1689341400,
-                              1689600600,
-                              1689687000,
-                              1689773400,
-                              1689859800,
-                              1689963697
-                         ]
-                    }
-               },
-
-          }
-     ]
-}
-
-export default stockData;
