+++ StockMarket_app/index.html
@@ -31,7 +31,7 @@
           <main class="workspace">
   
                <div class="search-bar">
-                   <select name="period" id="period-drpdwn">
+                   <select name="period" id="period-drpdwn" title="Period">
                          <option value="1mo" selected>1 Month</option>
                          <option value="3mo">3 Month</option>
                          <option value="1y"> 1 year</option>
@@ -59,6 +59,6 @@
           </footer>
      </div>
 
-     <script src="./index.js"></script>
+     <script type="module" src="./index.js"></script>
 </body>
 </html>
